finish navbar layout & add hover animations, fns work"

+++ src/components/Nav.js
@@ -5,16 +5,18 @@ import "../styles/nav.css";
 
 const Nav = ({ user: { loggedIn }, location: { pathname } }) => {
   const handleLogoutClick = (e) => {
-    localStorage.clear();
-    document.location.reload();
-    window.location.href = "/";
+    if (e) {
+      localStorage.clear();
+      document.location.reload();
+      window.location.href = "/";
+    }
   };
   const [burger, setBurger] = useState(false);
   return (
     <div className="navbar">
       <div className="left-nav btn-wrap">
         <div>LOGO</div>
-        <div>WATCHLIST</div>
+        <button className="nav-m-i btn">Watchlist</button>
       </div>
       <div className="right-nav btn-wrap">
         <Link className="nav-m-i btn" to="/profile">
@@ -22,9 +24,8 @@ const Nav = ({ user: { loggedIn }, location: { pathname } }) => {
         </Link>
         <button
           className="nav-m-i btn"
-          onClick={(e) => {
-            handleLogoutClick(e);
-          }}
+          styles={{ borderStyle: "none" }}
+          onClick={(e) => handleLogoutClick(e)}
         >
           Logout
         </button>
